Make the timeline 64-bit
This both gives us a little more room for shorter timesteps and allows more
than 1024 syncpoints/snapshots.

+++ libgadget/types.h
@@ -5,7 +5,7 @@
 
 /*Define some useful types*/
 
-typedef int32_t inttime_t;
+typedef int64_t inttime_t;
 
 typedef uint64_t MyIDType;
